Issue #3608 - fix merge of tests jetty-10
